+++ backend/services/storage_areas.service.js
@@ -8,6 +8,7 @@ const {
   processKeyValues,
   setUpdateOptions,
   setDeleteOptions,
+  filterImmutableData,
 } = require("../utils/serviceHelpers");
 
 const table = {
@@ -68,25 +69,29 @@ const createStorage_Area = async (storage_areaModel, storage_areaData) => {
 
   //If Primary key value is given, check if it is already in use.
   if (storage_areaData?.StorageAreaID) {
-    const exists = await storage_areaModel.findByPk(
-      storage_areaData.StorageAreaID,
-      {
-        raw: true,
+    const [instanceObj, isCreated] = await storage_areaModel.findOrCreate({
+      where: {
+        StorageAreaID: storage_areaData.StorageAreaID,
       },
-    );
+      defaults: storage_areaData,
+    });
 
-    if (exists) {
-      console.log(exists);
+    if (!isCreated) {
+      console.log(
+        `Storage Area With ID "${storage_areaData.StorageAreaID}" Already Exists!`,
+      );
       throw new AppError(
         `Storage Area With ID "${storage_areaData.StorageAreaID}" Already Exists!`,
-        200,
+        409,
       );
     }
+
+    return instanceObj.get({ plain: true });
   }
 
   const result = await storage_areaModel.create(storage_areaData); //{fields: []} to exclude injected key-values.
 
-  return result;
+  return result.get({ plain: true });
 };
 
 //useEmpty is an object with booleans used for flagging values that should be set to null.
@@ -107,15 +112,80 @@ const updateStorage_Areas = async (
     true,
   );
 
+  filterImmutableData(primaryKeyValuesArr, table.name);
+
+  if (primaryKeyValuesArr[0].StorageAreaID.length === 0) {
+    throw new AppError("Cannot modify starter data!", 400);
+  }
+
   const options = setUpdateOptions(table.primaryKeys, primaryKeyValuesArr);
 
-  const result = await storage_areaModel.update(storage_areaData, {
+  await storage_areaModel.update(storage_areaData, {
     where: options,
   });
 
-  //result[0] is just the number of affected rows.
-  //Sequelize doesn't return anything else for MySQL.
-  return "Updated rows: " + result[0];
+  const updatedRows = await storage_areaModel.findAll({
+    where: options,
+    raw: true,
+  });
+
+  if (!updatedRows || updatedRows.length === 0) {
+    throw new AppError(
+      "No record(s) found matching the provided identifier(s).",
+      404,
+    );
+  }
+
+  if (updatedRows.length >= 50) {
+    return null;
+  }
+
+  return updatedRows;
+};
+
+const patchStorage_Area = async (
+  storage_areaModel,
+  StorageAreaID,
+  patchData,
+) => {
+  delete patchData?.StorageAreaID;
+
+  if (!patchData || Object.keys(patchData).length === 0) {
+    throw new AppError("Body cannot be empty.", 400);
+  }
+
+  const dataToFilter = { StorageAreaID: [StorageAreaID] };
+
+  filterImmutableData(dataToFilter, table.name);
+
+  if (dataToFilter.StorageAreaID.length === 0) {
+    throw new AppError("Cannot modify starter data!", 400);
+  }
+
+  const result = await storage_areaModel.update(patchData, {
+    where: { StorageAreaID },
+  });
+
+  if (result[0] > 0) {
+    const updatedRow = await storage_areaModel.findOne({
+      where: { StorageAreaID },
+    });
+
+    return updatedRow.get({ plain: true });
+  }
+
+  const existingRow = await storage_areaModel.findOne({
+    where: { StorageAreaID },
+  });
+
+  if (!existingRow) {
+    throw new AppError(
+      `No record found for storage area: ${StorageAreaID}`,
+      404,
+    );
+  }
+
+  return existingRow.get({ plain: true });
 };
 
 const deleteStorage_Areas = async (storage_areaModel, storage_areaData) => {
@@ -126,20 +196,33 @@ const deleteStorage_Areas = async (storage_areaModel, storage_areaData) => {
     storage_areaData,
   );
 
-  const options = setDeleteOptions(keyValuesArr);
+  if (!keyValuesArr || keyValuesArr.length === 0) {
+    throw new AppError(
+      "Deletion Denied: No valid matching criteria provided.",
+      400,
+    );
+  }
+
+  const options = setDeleteOptions(keyValuesArr, "StorageAreaID");
 
   const result = await storage_areaModel.destroy({
     where: options,
   });
 
-  //Destroy() returns a number for how many rows where affected.
-  //Sequelize doesn't return anything else for MySQL.
-  return "Updated rows: " + result;
+  if (result === 0) {
+    throw new AppError("No matching records found to delete.", 404);
+  }
+
+  return {
+    deletedRows: result,
+    identifiers: storage_areaData,
+  };
 };
 
 module.exports = {
   getStorage_Areas,
   createStorage_Area,
   updateStorage_Areas,
+  patchStorage_Area,
   deleteStorage_Areas,
 };
